[SCRUM-150] Refactor : 게시물 수정 기능을 변경한다

+++ src/main/java/com/kkinikong/be/community/service/CommunityService.java
@@ -31,8 +31,10 @@ import com.kkinikong.be.community.domain.document.CommunityPostDocument;
 import com.kkinikong.be.community.domain.type.Category;
 import com.kkinikong.be.community.dto.request.CommunityCommentRequest;
 import com.kkinikong.be.community.dto.request.CommunityPostRequest;
+import com.kkinikong.be.community.dto.request.CommunityPostUpdateRequest;
 import com.kkinikong.be.community.dto.response.CommentListResponse;
 import com.kkinikong.be.community.dto.response.CommentResponse;
+import com.kkinikong.be.community.dto.response.CommunityPostImageResponse;
 import com.kkinikong.be.community.dto.response.CommunityPostInfoResponse;
 import com.kkinikong.be.community.dto.response.CommunityPostListResponse;
 import com.kkinikong.be.community.dto.response.CommunityPostPopularResponse;
@@ -92,13 +94,14 @@ public class CommunityService {
   }
 
   @Transactional
-  public void postCommunityPostImage(Long postId, List<MultipartFile> files, Long userId) {
-    if (files == null || files.isEmpty()) return;
-    if (files.size() > 3) {
+  public CommunityPostImageResponse postCommunityPostImage(
+      Long postId, List<MultipartFile> files, Long userId) {
+    if (files == null || files.isEmpty()) return new CommunityPostImageResponse(List.of());
+
+    List<CommunityPostImage> remainImageList =
+        communityPostImageRepository.findAllByCommunityPostId(postId);
+    if (remainImageList.size() + files.size() > 3) {
       throw new CommunityException(CommunityErrorCode.COMMUNITY_POST_IMAGE_SIZE_LIMIT);
-    }
-    if (communityPostImageRepository.existsByCommunityPostId(postId)) {
-      throw new CommunityException(CommunityErrorCode.COMMUNITY_POST_IMAGE_ALREADY_EXISTS);
     }
 
     CommunityPost communityPost = getCommunityPostOrThrow(postId);
@@ -110,7 +113,14 @@ public class CommunityService {
       communityPostImageRepository.save(
           CommunityPostImage.builder().communityPost(communityPost).imageUrl(url).build());
     }
-    communityPost.updateThumbnailUrl(imageUrl.get(0));
+
+    if (!remainImageList.isEmpty()) {
+      communityPost.updateThumbnailUrl(remainImageList.get(0).getImageUrl());
+    } else {
+      communityPost.updateThumbnailUrl(imageUrl.get(0));
+    }
+
+    return CommunityPostImageResponse.from(imageUrl);
   }
 
   @Transactional
@@ -350,18 +360,20 @@ public class CommunityService {
   }
 
   @Transactional
-  public void updateCommunityPost(Long postId, CommunityPostRequest request, Long userId) {
+  public void updateCommunityPost(Long postId, CommunityPostUpdateRequest request, Long userId) {
     CommunityPost communityPost = getCommunityPostOrThrow(postId);
     validatePostOwner(communityPost, userId);
+    List<String> remainingImageUrls = request.remainingImageUrls();
 
-    // 기존 이미지 모두 삭제
     communityPostImageRepository
         .findAllByCommunityPostId(postId)
         .forEach(
             image -> {
-              imageService.deleteFile(image.getImageUrl(), S3Bucket.COMMUNITY_POST_IMAGE);
+              if (!remainingImageUrls.contains(image.getImageUrl())) {
+                imageService.deleteFile(image.getImageUrl(), S3Bucket.COMMUNITY_POST_IMAGE);
+                communityPostImageRepository.deleteById(image.getId());
+              }
             });
-    communityPostImageRepository.deleteAllByCommunityPostId(postId);
 
     communityPost.update(request.title(), request.content(), request.category());
